# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API",
    "version": "1.0.2"
  },
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "marca": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        }
      },
      "required": ["marca", "modelo"]
    },
    "Concesionarios": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "coches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coche"
          }
        }
      },
      "required": ["nombre", "direccion"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Get all concesionarios.",
        "responses": {
          "200": {
            "description": "Array with concesionarios.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionarios"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create concesionarios",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "Concesionarios",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Get one concesionarios by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch one concesionarios.",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        }
      },
      "put": {
        "summary": "Update one Concesionarios by Id.",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "Concesionarios",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      },
      "delete": {
        "summary": "Delete one Concesionarios by Id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Get all coches for a concesionario by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Array with coches.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionarios"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a coche for a concesionario by ID.",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "Coches",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    },
    "/concesionarios/{id}/coches/{idCoche}": {
      "get": {
        "summary": "Get one coche for a concisionario by Id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "idCoche",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "One coche.",
            "schema": {
              "type": "string",
              "$ref": "#/definitions/Coche"
            }
          }
        }
      },
      "put": {
        "summary": "Update one coche for a concisionario by Id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "idCoche",
            "type": "integer",
            "required": true
          },
          {
            "in": "body",
            "name": "Coches",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      },
      "delete": {
        "summary": "Delete one coche for a concisionario by Id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "idCoche",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    }
  }
}